Abandon unserviceable AWS access
Discontinuation of using AWS as an alternative download source for .jp2
previews and single images (Image Selection mode).
Since public access on the AWS bucket was disabled on 7th of August
2018, now generally use DHuS.
As a consequence, some tasks perform slower now due to the usually poor
response times of ESA's OData producer.